Update README with additional functionality details
A script that will return the old top bar from 2017-2019. Function: Replace the new bar permanently, as the script can turn it on and off (I will soon make it so that it saves the user's choice). It can change the cursor and return the old menu, but I have not yet found a way to disable it and replace it with this script. It also removes unnecessary CoreGui, which can overload your device.